pyproject.toml
[project]
name = "chromey"
version = "0.2.0"
description = "Chromey: a Chrome-only browser-use proxy with a sidepanel extension and LM Studio integration."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "browser-use>=0.12.1",
  "PyYAML>=6.0.2",
  "starlette>=0.48.0",
  "uvicorn>=0.37.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/chromey"]
